requirements-ai.txt
# Optional: enable AI mode
openai>=1.40.0
PyYAML>=6.0.1
python-hcl2>=4.3.5
diagrams>=0.23.4
